requirements.txt
flask==3.0.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
gunicorn==21.2.0

